Added four duplicate image links and a JavaScript function to redirect to 'issue1/index.html' on image click
This diff adds four duplicate image links with the same source and alt attribute as the existing image link. It also adds a JavaScript function called 'redirectTo' that changes the current webpage to the URL passed as an argument when called. The function is used as an onclick event handler for the first new image link.