Update magazine CSS width and HTML links for additional editions
The CSS file for the magazine has been updated to change the width of the body from 400px to 922px. The HTML file has been updated to include links for two additional magazine editions and removed the "About" link. The page count in the JS file has been updated to 26 to reflect the new number of pages.